% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_n_clones.R
\name{est_n_clones}
\alias{est_n_clones}
\title{est_n_clones}
\usage{
est_n_clones(
  ctrl_muts,
  cond_muts,
  range_ctrl = 2:8,
  range_cond = 2:8,
  n_repeat = 100
)
}
\arguments{
\item{ctrl_muts}{a numeric vector containing the number of mutations found in individual control clones.
At least 3 values are needed.}

\item{cond_muts}{similar to \code{ctrl_muts}, but for exposed clones}

\item{range_ctrl}{a numeric vector containing the numbers of potential control clones to be tested.
Default: \code{2:8}}

\item{range_cond}{similar to \code{range_ctrl}, but for exposed clones}

\item{n_repeat}{the number of iterations of tests on which the final P-value is based.
At least 100 is recommended. A higher number gives a more reliable result.}
}
\value{
a heatmap plot indicating the number of control and exposed clones to sequence to determine a statistical significant increase.
}
\description{
estimate the number of control and exposed clones needed for a significant result
based on the number of mutations in a few inital cells
}
